New map array split
Rearrange map data storage. Instead of an 8-byte struct array _m
for most of the data and an auxiliary byte array _me to hold m7,
use a byte array _mth to only hold the type and height of a tile
(and eventually only height and tropic zone) and an 8-byte struct
array _mc for the tile contents (minus its type, for now).