New function GetTileProcs
There is a global variable _tile_type_procs that holds an array
of pointers to the tile procedure table for each tile type.
When such a procedure is needed for a tile, its type is used as
an index into this array to reach the corresponding table.
Encapsulate this access into a new function GetTileProcs, which
returns a pointer to the procedure table for a given tile,
so most of the code do not need to know anything about how
these procedure tables are stored or retrieved. This abstraction
will be useful when switching to the new map array.

(As a side effect, we now have to make DrawTile_Void non-static,
because it is required to draw off-map parts of a viewport.)