Add old tile type definitions in AfterLoadMap
When the new map array is implemented, the definitions for
the current tile types will be removed, but they will still
be required in AfterLoadMap, to convert old savegames.
Preemptively copy the tile type enum, under a different name,
for use in AfterLoadMap, so it remains available after the
current tile types no longer exist in the rest of the code.