Make FixOldMapArray independent of current tile definitions
FixOldMapArray is in charge of converting old map arrays (TTO/TTD/TTDP)
into the current format (or, rather, into something that AfterLoadGame
can fathom) However, the interpretation of the old map array must not
depend on the current definitions used for the actual, current map
array (as GetTileType), because these can change between versions
but the meaning of the bits in an old map array cannot. So, make
FixOldMapArray, and a few other places in oldloader_sl, use its own
definitions instead of relying on the global ones.