Make smallmap code independent of map array tile types
Disentangle smallmap code from the particular tile type definitions
used in the map array. By defining a new enum SmallmapTileType and
using it in the smallmap code, we can stop depending on a particular
map array implementation. The code becomes shorter and cleaner, too,
as we remove the tiletype "importance" array, because smallmap tile
types are already sorted by "importance".