Rework drawing of signals
Rework DrawSingleSignal, so that it does all the required
computations for actually drawing a signal. As it was,
DrawSingleSignal accepted several parameters: track, signal
condition, image sprite and a special index to fetch the
signal position from an internal array. Instead, now it
takes just a trackdir argument, and retrieves said data
from an extended array, which holds signal position, sprite
offset and bit to test for signal presence/condition.