New function IsTileFlat
There are many places in the code where GetTileSlope is called
on a tile just to check its result against SLOPE_FLAT; introduce
a new function IsTileFlat to encapsulate this check. This improves
code readability but, more importantly, improves performance:
GetTileSlope does a complex dance to accurately compute the tile
slope, which can be skipped if we are only interested in whether
the tile is flat or not; also, we can bail out earlier, as soon
as two corners have different heights.