Store a trackdir instead of a track bitmask for trains and ships
Trains and ships store the track they are driving on in their
struct. However, they store it as a *bitmask*, where only one
bit is set. (And then there are special values for wormholes,
depots, etc.) This is a non-intuitive way of doing things, to
say the least, and it implies using FindFirstTrack or similar
tricks to retrieve the (single) bit set in the mask whenever
it is needed.

Change the code to stop storing the track as a bitmask. While
at it, store a trackdir instead of a track, because sometimes
it is the trackdir that is needed, and getting a track from a
trackdir is immediate but the other way around is not.