Rework wormhole enter code
Vehicles will now acquire wormhole status right when they
leave the wormhole head into the middle part, not on the last
bridgehead frame or at some random intermediate point in a
tunnelhead. (In particular, this decouples hiding a vehicle
when entering a tunnel from declaring it in the wormhole.)
This unifies entry behaviour between bridges and tunnels,
removes one of the reasons why vehicle enter procs exist
and is required for custom bridgeheads, when we will not be
able to assume that a vehicle cannot change direction after
it has entered a bridgehead until it reaches the other end.