Introduce some sanity in the handling of train depots
_fractcoords_{behind,enter} hold the per-direction intra-tile
coordinates at which wagons enter/exit depots, with the x
coordinate in the lower 4 bits and the y coordinate in the
higher 4 bits, instead of using one byte each. Since saving
8 bytes of read-only data section is not worth the added code
complexity, just split them into _x and _y arrays.

Also, _deltacoord_leaveoffset is encoded as an array of size 8,
when it should really be two arrays of size 4, one for x and
another for y.