New function ClearRoadType to clear a roadtype from a tile
Add a new function ClearRoadType to clear all road bits for a
certain roadtype (road or tram) from a tile. It takes care of
assigning the owner of the tile to the owner of the other roadtype,
if there is one, or of fully clearing the tile if it was the last
remaining roadtype. At present this is only useful in a couple of
places, but more will come with extended road bridgeheads.