New struct PFPos for a pathfinder position (track and trackdir)
Add a new struct PFPos to stand for a pathfinder position (track
and trackdir currently, but this will change when a wormhole's
middle part has to be considered in pathfinding). This will serve
as an abstraction of the many places where a (track, trackdir)
pair is used to represent the position of a vehicle.