Rework EnsureNoTrainOnTrackBits
EnsureNoTrainOnTrackBits is the function called to check if there
is a train on a tile that would interfere with a given set of
tracks, which is always unless we are only interested in one
of the short tracks and the train is on the parallel one.
This means that there is at most one track on which a train can
be and not cause the check to fail. Therefore, it is better to
pre-compute this allowable track, if any, and use it in the
callback by just comparing it with the track that the given
train is running on, instead of doing expensive overlap checks.