Switch GetOtherTunnelEnd to a faster implementation
To find the other end of a tunnel, we do not need to repeatedly
check whether tiles along the way are tunnel tiles and head the
right direction; the tunnel must necessarily end at the first
tile whose height is the same as the tunnel base height.
So just check for this, which is faster.