Rearrange updating of signals in the train controller
Updating signals in the train controller must happen under the
following circumstances:

- When the first wagon (locomotive) enters a tile, signals must
be updated on the segment containing the new tile.

- When the last wagon leaves a tile, signals must be updated
on the segment containing the old tile.

Previously, all updates were done at the end of the main controller
loop, but this relied on the fact that the segment containing the
old tile was easily computed even when the last wagon had already
moved. When signals on bridgeheads are implemented, this will no
longer be the case, so make the necessary adjustments in advance.
Now, the segment to be updated when the last wagon leaves a tile
is added to the signal buffer *before* the wagon moves, when it is
easy; updating, however, has to be defered until it has actually
moved (otherwise, the not-yet-moved wagon itself will make the
segment busy). Moreover, updating the segment just left has to
come before updating the segment just entered, when the last wagon
is also the first (a consist of exactly one locomotive).