New function FindUnpoweredReservationTrain
In CmdConvertRail, there are two places where the code checks if
a rail conversion makes a train's reservation invalid because
it would now go through a railtype on which the reserving train
is unpowered, one in the tunnelbridge path and another in the
non-tunnelbridge path. Make a separate function for this check
and use it at both places.