New function CheckWaitingPosition to analyse a waiting position
There is a lot of duplicated code in IsSafeWaitingPosition and
IsWaitingPositionFree, while their roles are closely related.
Worst of all, they are sometimes called in succession (all uses
of newly introduced IsFreeSafeWaitingPosition), which means
that some checks are run twice, first to see if a position is
safe and then to see if it is free.

To solve this, we introduce a new function CheckWaitingPosition
that merges the functionality of IsSafeWaitingPosition and
IsWaitingPositionFree. It has four modes of operation, selected
by parameter cb:

- When cb is PBS_CHECK_FULL, it performs a full analysis of the
given position; it returns PBS_UNSAFE if the position is not
safe, PBS_BUSY if the position is safe but not free and PBS_FREE
if it is safe and free.

- When cb is PBS_CHECK_SAFE, it only checks whether the given
position is safe; it returns PBS_UNSAFE if and only if it is
not safe, but the return value is unspecified between PBS_BUSY
and PBS_FREE if it is safe. This behaviour is encapsulated by
function IsSafeWaitingPosition.

- When cb is PBS_CHECK_FREE, the function assumes that the
given position is safe (it should have been checked before, or
otherwise known to be safe) and checks whether it is free; it
returns PBS_FREE if and only if it is free, but the return
value is unspecified between PBS_UNSAFE and PBS_BUSY if it is
not. Also, the return value is undefined if the position is
actually unsafe. This behaviour is encapsulated by function
IsWaitingPositionFree.

- When cb is PBS_CHECK_SAFE_FREE, the function checks whether
the given position is both safe and free; it returns PBS_FREE
if and only if it is, but the return value is unspecified
between PBS_UNSAFE and PBS_BUSY if it is not. This behaviour
is encapsulated by function IsFreeSafeWaitingPosition.