New function GetAvailShipTrackdirs
Add a new function GetAvailShipTrackdirs in ship_cmd.cpp,
to call GetTileTrackStatus on a tile and mask out trackdirs
not reachable from a tile side. This function effectively
replaces GetTileShipTrackStatus and GetAvailShipTracks.

(Currently all uses of GetAvailShipTrackdirs are immediately
composed with TrackdirBitsToTrackBits, so this change may
seem pointless, but it will become useful in the next commit.)