Remove RVS_DRIVE_SIDE and simplify road vehicle drive data
Reserving a bit in RoadVehicleStates just to access drive data
is a terrible hack that must go away. To this end, we repack
the road vehicle state bits and we rearrange drive data into a
two-dimensional array where the first dimension is the drive side
(xor overtaking status). This also means that the overtaking
field in a road vehicle struct now takes sane values (0 or 1).