Switch trams to straight drive data after a long turn
Handling of tram reversing in the road vehicle controller is
somewhat convoluted. This is partly due to the fact that trams
can take two different types of turn: a normal big turn when
there is only one tram roadbit on the tile; and a short turn
when the road ends and the tram would otherwise get stuck.

This is handled as follows: The tram road drive data array for
reversing trackdirs contains the drive entries for a long turn.
The short turn is a subsegment of this array, so trams doing a
short turn directly jump to a frame in the middle of the array,
and then there is a special marker later that signals the end
of the short turn, which tells the controller to call the
pathfinder and not continue with the long turn data.

However, it turns out that once that a tram is in the middle of
turning (in the common part between short and long turns), it
does not really matter whether the turn started as short or
long, since the result will be the same. In a short turn, the
tram will be switched to a different drive data array when it
reaches the short-turn-end marker; in a long turn, it will keep
going with the long turn data, but the drive entries after that
marker are exactly the same as those in the straight drive data
array. As a result, we can remove the entries after the
short-turn-end marker in the data array for reversing
trackdirs, if we make trams doing a long turn switch to the
straight drive data at that point.