Improve detection of tram long turns
When deciding if a tram is doing a long turn, the code checks
for tiles with exactly one roadbit, excluding depots. However,
only normal road tiles are suitable for long turns, as they
are the only ones that can have one roadbit (other than depots),
so just check for that. As a bonus, we no longer have to go
through the catch-all GetAnyRoadBits, but can call GetRoadBits
directly.