Simplify return value from TrainEnterTile
The only remaining possible return values from TrainEnterTile are
VETSB_CONTINUE, which is the normal case, and VETSB_ENTERED_STATION,
when entering a station. Instead of using VehicleEnterTileStatus as
return type from TrainEnterTile, just use StationID, and return
the station id when actually entering a station and INVALID_STATION
otherwise.