New function SaveDumper::Dump
Add a new function SaveDumper::Dump to dump the contents of a
SaveDumper into another one.

The goal of this function is to allow callers to make a partial
dump of contents into a temporary SaveDumper, then add it to the
main dumper. Since the savegame structure forces RIFF chunks and
(sparse) array elements to have their size stored before their
contents, this new function is useful when said size is not
known until after the data have been processed, which is the
case with some handlers (see SlArrayAutoElement and the
NL_CALCLENGTH madness, for instance) that are currently forced
to run through their data twice, once for computing their length
and again to actually store them. With this new function, they
will ultimately be able to stack their data in a temporary
dumper in a single pass and then dump its size and contents
to the savegame.