Script save functions now take a SaveDumper argument
Pass the save dumper used for saving also to the script saving
code, so it does not have to rely on global functions either.

Since script save functions were only called from chunk handlers
through SlArrayAutoElement, which calls its proc argument twice,
once to compute the element length and another one to actually
save the data, and this behaviour depends on special checks in
SlObject, which is no longer called from script functions, the
use of SlArrayAutoElement is also removed. Instead, a temporary
dumper is set up to hold script data as it is generated, and
then it is sent to the main save dumper when finished.