Pass the save dumper down to the chunk handlers
Pass the save dumper used for saving down to the chunk handlers,
so that they do not depend on global functions to store the data,
but have access to SaveDumper methods directly. This will also
allow for further improvements down the road.

Note that some chunk handlers still depend on global saveload
functions; this will be fixed in later commits.