Rearrange SaveLoad struct, and add new subtype StrType
The enum VarTypes encodes conversion of saveload data between
memory and file. However, not all pairings between memory type
and file type are possible, and there are further restrictions
depending on the saveload type. Specifically, when type is
SL_STR, the file type must necessarily be SLE_FILE_STRING,
and it must be paired with one of the four SLE_VAR_STR* memory
types; on the other hand, when type is not SL_STR, the file
type for the conversion cannot be SLE_FILE_STRING, and the
memory type cannot be any of the SLE_VAR_STR* values.
Additionally, the special flags SLS_ALLOW_* only make sense
with SL_STR.

This means that using the same VarTypes enum for both SL_STR
and SL_VAR/SL_ARR is not a good choice. Instead, we remove
SLE_FILE_STRING, SLE_VAR_STR* and SLS_* from VarTypes and move
them into a new enum StrTypes. With SL_VAR/SL_ARR, we keep on
using the old VarTypes enum; with SL_STR, we switch to the new
StrTypes enum.

With this change, the conversion field in SaveLoad has the
following meaning (the change only affects SL_STR):

- when type is SL_VAR or SL_ARR, it holds a VarType, which encodes
the required conversion between memory and file;

- when type is SL_STR, it holds a StrType, which also encodes the
required conversion between memory and file, but which is different
from VarType;

- when type is SL_REF or SL_LST, it holds the reference type;

- when type is SL_WRITEBYTE, it holds the byte to write;

- when type is anything else, it is unused.