Fix usage of GetVariableAddress in SlObjectMember
When using SL_*_INCLUDE, the pointer passed recursively to SlObject
should be the base object pointer, not a non-existing variable
pointer, because SL_*_INCLUDE members do not define a variable
address (cfr. SlCalcObjMemberLength). It happens to work because
all SL_*_INCLUDE macros have a NULL address, which translates as
(ptrdiff_t)0, so ptr==object, but this is a kludge and should go
away.