Unify includes of SaveLoad objects
Instead of having a dedicated SL_* code for each possible inclusion
of a SaveLoad object within another, have a single SL_INCLUDE value
and use the address field in the SaveLoad description to point to
the included SaveLoad array. This is more future-proof and makes
code simpler, because SlObject and friends do not have to know
about different include codes or to fetch the included description.