Special-case SLE_VAR_NAME in LoadBuffer::ReadVar
SLE_VAR_NAME is only used in combination with SLE_FILE_STRINGID,
and only in a few saveload entries for strings that were saved as
integers in past savegame versions but are now stored as proper
strings in memory and in the savegame. Since its only use is in
converting old savegames, deal with it in LoadBuffer::ReadVar,
so that WriteValue does not need to know anything about
SLE_VAR_NAME or CopyFromOldName, making it clear that it is
a savegame compatibility hack.