Remove reference in LZOLoadFilter to _sl_version
The LZO load filter treats savegame version 0 specially, because
in that version block sizes were not saved endian-safely. To make
up for this, the read method checks the current savegame version
and skips endian correction in version 0. To remove the last
reference to the global _sl_version object, make this check depend
on a boolean variable set in the filter during construction,
instead of accessing _sl_version every time.