Remove fields sf and dumper from SaveLoadParams
Declare a local SaveDumper in the function that needs it,
instead of having a permanent pointer in _sl. Also, do not
store the save filter in _sl.

Note that SaveFileToDisk takes over ownership of the writer and
the dumper that it receives, and will delete them when finished,
so the caller must not do so itself. This is because, in threaded
saves, the caller cannot know when they can be deleted.