Remove SLEG_CONDNULL, and make SLE_CONDNULL global
Remove SLEG_CONDNULL, which is never used, and change the
definition of SLE_CONDNULL to set the global flag, so no
auxiliary struct is required; since SLE_VAR_NULL means that
the value will always be discarded, the address should never
be taken, and this is clearer.

While at it, fix a stray SLEG_NULL, which is undefined.