Rearrange GetSaveLoadErrorString and its callers
Currently, GetSaveLoadErrorString returns a (pointer to a static)
string with a description of the last saveload error. This string
is governed by either STR_ERROR_GAME_SAVE_FAILED or
STR_ERROR_GAME_LOAD_FAILED, both of which end with a newline
and a {STRING} that hooks into the actual saveload error (and
which can, in turn, have additional parameters).

Callers of GetSaveLoadErrorString, most of them in openttd.cpp,
however, always make use of its return value in the same fashion,
namely as first argument in a call to ShowErrorMessage() (via
STR_JUST_RAW_STRING), in which the second argument is always
INVALID_STRING_ID. This means that ShowErrorMessage()'s built-in
multiline capability is ignored, by forcing an artifical newline
in the first string but leaving the second one empty.

To improve the situation, we change the behaviour of
GetSaveLoadErrorString to only return the actual saveload error
(without STR_ERROR_GAME_SAVE_FAILED/STR_ERROR_GAME_LOAD_FAILED).
This string will in turn be used as _second_ argument to
ShowErrorMessage, so it will naturally appear on a different
line. The first line will be STR_ERROR_GAME_SAVE_FAILED/
STR_ERROR_GAME_LOAD_FAILED, now stripped of their trailing
newline and {STRING} hook; this means that callers of
GetSaveLoadErrorString must now be aware of whether the
last error was triggered by a save or a load, but they all do.
This achieves the same result as before, with less effort.