Make LoadOldSaveGame return errors consistently
LoadOldSaveGame is not consistent in returning errors: the global
saveload error is set (through SetSaveLoadError) if a savegame
fails the data integrity check, but not if the file cannot be
opened at all (for instance, due to lack of permissions). Fix
this by passing a pointer to a SlErrorData struct through which
LoadOldSaveGame can return an error (and avoid one use of
SetSaveLoadError in the process).