Merge file opening in LoadGame
Merge file opening in LoadGame between old and non-old savegames.

(This introduces a slight change in behaviour: When a file to be
opened as an old savegame cannot be read, the error message will
be in category 'sl', instead of 'oldloader'. Which is probably
better anyway, because it means there is a problem with the
file itself, not with its contents.)