Handle signals in LoadWithFilterMode
Error handling in AfterLoadGame is a mess. Some error conditions
trigger resetting of signal handlers and returning false, while
others throw an exception. However, returning false will leave
the caller (DoLoad) without an error description unless
SetSaveLoadError is used (and sometimes it is not), and throwing
an exception will not reset the signal handlers.

This commit fixes the latter. Instead of having AfterLoadGame set
and reset signal handlers, it is now done in LoadWithFilterMode,
around the try-catch block; after all, if their goal is to catch
savegame corruption, they should be in effect during the whole
process, not just during AfterLoadGame (a corrupted savegame can
trigger a segfault well before AfterLoadGame runs). This means
that all the relevant signal-related functions are now in
saveload.cpp.