New functions to add update points to the signal buffer
Add new functions AddDepotToSignalBuffer, AddBridgeToSignalBuffer
and AddTunnelToSignalBuffer to add the corresponding update points
to the signal buffer, so that AddSideToSignalBuffer is never called
with INVALID_DIAGDIR as side. Currently all three functions just
behave as AddSideToSignalBuffer with INVALID_DIAGDIR, but this is
about to change, and this commit prepares the call sites in advance.