Add support for experimental savegames
Add a global boolean flag to set experimental status for savegames.
Savegames made with a binary with this flag set will have a special
value in their version field, and are only guaranteed to load in
a binary made from the same sources, making them effectively
unversioned, without any backward or forward compatibility whatsoever.
This should be used as an aid during phases of development where the
savegame format is expected to change many times in a short burst, to
avoid too many savegame version bumps and, more importantly, savegame
compatibility code creep.