Move some tile-related types to a new file tile/type.h
This is the first commit in a series to more tile-related data
and functions under a new subdirectory src/tile. The goal will
be to have a copy of most of the functions in *_map.h files,
operating on a tile pointer instead of on a TileIndex, under
src/tile. This is done mainly for modularity, but it can come
in handy if we ever implement some sort of clipboard.