Ignore dotless includes in depend
Ignore dotless includes in depend, such as '#include <map>'. The
current implementation of depend treates those includes normally,
but they are always system includes that should not be found in
the source tree, and if a directory named 'map' happens to exist
under objs/release (as will soon be the case), depend will output
its path as a dependency, and then make will guess it is a binary
to be linked from 'map.o' and get really confused.

Of course, the right thing to do here would be for depend not to
track includes using <...>, since they are supposed to be system
includes, not files in our own tree. However, some of the files
under src seem to be using <...> to include sibling headers, and
they would have to be fixed first. So, this workaround should do
the trick for now.

(One wonders why a custom dependency generation tool is needed
at all, especially one that does not work properly, but that is
a story for another time.)