Move fix for vehicles on a bridge in AfterLoadGame
In savegame version 42, ground vehicles on a bridge switched
from storing the virtual tile they were on to storing one of the
bridgeheads together with a special wormhole value as trackdir.
However, the code in AfterLoadGame that accounts for this change
runs after AfterLoadVehicles, which indirectly calls a variety
of map accessors, for instance to get the rail type a train is
running on. If a train has not had its tile field fixed yet, the
code will try to get the rail type of the tile under the bridge,
which may not have any rail at all, leading to bogus values.
Fix this by making the conversion for old savegames run before
AfterLoadVehicles.