Road tunnel/bridge heads have no trackbits wrt catenary drawing
Road tunnels and bridges do not have any rail, so do not treat them
as if they had when drawing a catenary. The code did not crash just
because there is no check in GetRailType for the tile to actually
have any rail, and worked because the relevant bits in the map array
for road tunnels and bridges were always 0, and rail type 0 does not
(usually) have a catenary, but happening to work is not a valid
excuse for wrong code.