Handle end of line consistently in the track follower
The track follower handled end of line differently depending
on how the end of line was detected; a road that ended on a
clear tile was considered for reversing, but one that ended
against a depot facing the wrong way was not. This is fixed
now. (This only affected non-tram road vehicles.)