Use GetRailType instead of GetTileRailType in several places
Replace GetTileRailType by GetRailType in several places where
it can be proven that there is rail:

- GetTrainForReservation should only be called for reserved tracks;
therefore, it can assume that there is a railway track at its tile
and track arguments, and can safely call GetRailType instead of
GetTileRailType. (Moreover, if GetTileRailType were to return
INVALID_RAILTYPE, GetRailTypeInfo would immediately assert.)

- GetTileRailwayStatus only returns non-null for railway tiles,
level crossings and rail depots, tunnels and station tiles; and,
for all of these, GetTileRailType just calls GetRailType. Thus,
we can directly call GetRailType on tiles for which a previous
call to GetTileRailwayStatus returned non-null. This applies to
the track follower, the train controller and the catenary wire
masking function.