Remove num_dir_cycles from CmdBuildSingleSignal
The only user of num_dir_cycles in CmdBuildSingleSignal was the
scripting code for rail, to select the present signals to set in a
newly-built signal pair. By switching this code to use SIGNALS_COPY
instead of SIGNALS_BUILD, we can select those present signals in
the copy_data field, making num_dir_cycles unnecessary.