Move object counting out of Ptrs_OBJS
The goal of Ptrs_ functions should be to fix references; they should
not try to do other afterload tasks. Move object counting, and
removal of stale objects, out of Ptrs_OBJS and into a new function
AfterLoadObjects, to be called from AfterLoadGame, so that at the
time of calling AfterLoadMap has already run, unlike in Ptrs_OBJS,
and map accessors are guaranteed to work.