Invert road works count direction
Make the tile road works counter go downwards instead of upwards.
When road works start on a tile, the counter is set to 15, and is
then periodically decreased until it reaches 0. This way, we no
longer need special roadside values to flag road works in progress;
checking if the counter is not zero is enough.