Split catenary config array into two variables
Index 0 of the catenary config array in DrawCatenaryRailway is used
throughout the function, but index 1 is only used in the neighbour
loop, and values from one iteration are not used in the following
one. Make each index a different variable, and restrict the scope
of the neigbour config to the neighbour loop.