Improve packing of catenary sprite data
All catenary sprite data for a given track (and uphill/downhill
status) shared the same values, except for the sprite to use.
Instead of duplicating this information, define a new struct
to hold the sprite position and size (which is shared) and
a list of sprites to use depending on which pylons are present.