Add per-direction tile diffs to the map size structure
Add the per-direction tile differences between a tile and its
neighbours to the map size structure. These are used very often,
through TileOffsByDir and TileOffsByDiagDir, and now they can be
directly fetched from a global address. Previously, the coordinate
difference had to be fetched first, and then converted to a tile
index difference, which required an additional fetch of the map
X size, so we are halving the global accesses made.