Omit call to FindTrainOnPathEnd in FollowTrainReservation
In FollowTrainReservation, if the starting position is not reserved,
then ftoti.pos is set to the position of v itself, and then the
train that FindTrainOnPathEnd will find on that position is v.
As the only caller of FollowTrainReservation that supplies a
train_on_res pointer ignores the value received there if it is
precisely v, just skip calling FindTrainOnPathEnd in that case.