Switch ExtendTrainReservation to a ternary return value
Introduce an enum for the return value of ExtendTrainReservation,
with three values meaning that the path could be extended to a safe
position; that a choice or possible target tile was found; or that
the path could not be extended. ExtendTrainReservation now also
takes an extra PFPos parameter to store the first unsafe tile in
the second case.