Add VehicleTileIterator
Add a new iterator VehicleTileIterator for vehicles at a tile
(on the tile itself or in a depot or wormhole ending on the tile),
with the goal of replacing FindVehicleOnPos and HasVehicleOnPos.
Profiling shows that using the iterator is (unsurprisingly) faster
that calling {Find,Has}VehicleOnPos and having it call back through
a function pointer. Also, it allows for the callback code to be
inlined where it is needed, which makes for cleaner code. The
following commits will convert users of FindVehicleOnPos to the
new iterator.