Use VehicleTileIterator in CheckTrainCollision
Use VehicleTileIterator to check for train collisions in
wormholes in CheckTrainCollision. This requires a new function
FindTrainCollision to do the work of FindTrainCollideEnum,
and making the latter a wrapper for the former, because
FindTrainCollideEnum is also used as a callback for
FindVehicleOnPosXY, which does not have an equivalent iterator
yet.