Remove unreachable code in the road vehicle controller
When the road vehicle controller has to move a vehicle into a
new tile, it calls RoadVehEnterTile and checks if it returns
VETS_CANNOT_ENTER, in which case it either stops the vehicle (on
non-wormhole tiles) or attempts to turn it around (on wormhole
tiles). However, RoadVehEnterTile can return VETS_CANNOT_ENTER
only on road stop tiles, so remove the unreachable branch.