Compute distance for a nearby depot in FindClosestTrainDepot
Have FindClosestTrainDepot take a boolean parameter to select
whether to only look for nearby depots, instead of an integer
maximum penalty for the pathfinder. Since this penalty, if any,
depends on pathfinding settings, it is best left to compute to
FindClosestTrainDepot, which already has a switch on the active
pathfinder.