Compute distance for a nearby depot in FindClosestRoadDepot
Have FindClosestRoadDepot take a boolean parameter to select
whether to only look for nearby depots, instead of an integer
maximum penalty for the pathfinder. Since this penalty, if any,
depends on pathfinding settings, it is best left to compute to
FindClosestRoadDepot, which already has a switch on the active
pathfinder.