Stop using FindDepotData for road vehicles
Now that FindDepotData no longer has a best_length field, and
since road vehicles cannot reverse when looking for a depot,
a tile is all that is left for the road depot finder functions.
Remove the FindDepotData parameters from all of them, and make
them return the depot found as their normal return value.