Fix cost of end-of-line after a waypoint in pathfinders
Switching to FollowNext() introduced a bug in the rail node cost
functions when end-of-line was found after a waypoint (but without
any intervening junction), because the wrong tile would be tested
for being a free waiting position. Fix this, and make the surronding
code that deals with node cost after a waypoint slightly clearer.