Remove PBSTileInfo and introduce PFResult
The only places that still used struct PBSTileInfo were the train
pathfinding functions. Since all of them also had a parameter to
return whether a path was actually found or only guessed, grow
PBSTileInfo with a boolean flag into a new struct PFResult, which
now holds all of the extra information that the rail pathfinders
return.