Rework use of best_trackdir in ChooseTrainTrack
Use a new trackdir variable single_trackdir in ChooseTrainTrack if
the original tile only has one available trackdir, to make the code
easier to follow. This also fixes a bug where ExtendTrainReservation
would extend the reservation back to the original tile, but on a
different side, because the track has a loop, and ChooseTrainTrack
would get really confused, because it only checked if the tile was
the same.