Fix check for road stops in the road vehicle controller
Normally, the movement tables for road vehicles do not prescribe
a turn of more than 45 degrees for a vehicle, but there is a
notable exception. When a road vehicle reaches the stopping frame
of a standard stop, it has to make a full turn around. To allow
for this, the road vehicle controller inserted a delay for most
(but not all) changes of direction in the vehicle. However, this
is unnecessarily conservative, as it is enough to insert the delay
precisely at the stopping frame of a standard road stop. Just do
this, and document why this is needed.