Restrict overzealous assert in FollowPreviousRoadVehicle
In 4cb04776c041df2923f72c06264cfec4bcb086ac we extended the check
in FollowPreviousRoadVehicle that the new tile has the right road
bits for the chosen trackdir. However, if the new tile has no road
bits at all, the previous vehicle will have made a short turn using
a reversing trackdir, remaining in the old tile, in which case the
assert is wrong. Restrict the check to non-reversing trackdirs.