Redo handling of a possible signal in ChooseTrainTrack
ChooseTrainTrack is responsible for turning a path signal green
when extending a reservation across it. However, this was done by
unconditionally turning it green at first and then turning it back
red under some circumstances if reservation failed. Instead, just
record if there was a signal at the beginning and turn it green
later when really needed.