Introduce PathTile
Introduce PathTile in pathfinder_type.h to hold a pathfinder tile,
which can be either a map tile or a virtual tile in a wormhole,
and make PathPos derive from it. This is required to fix a bug in
YAPF node key comparison functions, but will come in handy for
other purposes as well.