Fix YAPF node key comparison functions wrt wormholes
Fix YAPF node key comparison functions for nodes in wormholes;
the wormhole status was not considered when comparing keys, so
it was theoretically possible for two nodes to compare equal
even if one was in a wormhole and the other was not. In practice,
however, their hashes would have been different, and the track
follower only generates positions in wormholes for the last
virtual tile in the wormhole, so the only way to generate a
collision would have been to place a railroad track heading into
a wormhole head from the wrong side, which would be a dead end
anyway. Still, it is better to have this fixed for correctness
and in case the track follower changes behaviour.