Move YAPF pathfinding to Astar
Move the core pathfinding loop in CYapfBaseT to a new method
in Astar, which takes the function to use to follow nodes as a
template argument. This makes Astar a generic implementation of
an A-star pathfinder algorithm, depending only on a Node struct
for path nodes and a follower function to find neighbours.