Do a static cast when calling follow in Astar::FindPath
Astar::FindPath is likely to be called with a follow function
that does not take the Astar struct itself as first argument,
but a derived class, since a pathfinder will often want to add
custom functionality to Astar. Make the call to follow perform a
static cast on this to the expected type, to make things easier
for the caller.