Remove CFollowTrackRailT::MaskReservedTracks
CFollowTrackRailT::MaskReservedTracks was a helper method that
checked if template parameter Tmask_reserved_tracks was true and
then dispatched the call to the underlying MaskReservedTracks
method in CFollowTrackRailBase if it was, or did nothing if it
was not. However, CFollowTrackRailT::MaskReservedTracks was only
ever called for specialisations of CFollowTrackRailT for which
Tmask_reserved_tracks was true, so removing it achieves the same
purpose by letting the call go to the base method directly.