Thoroughly rearrange rail YAPF
Define a new class template CYapfRailBaseT that only depends on
the YAPF node class and the track follower used, but not on the
destination or follower modules, and move everything in CYapfRailT
that does not depend on anything else there. This greatly reduces
the number of specialisations required for those functions.