Ignore wormholes in CYapfNodeKeyExitDir hash function
Ignore whether the current position is in a wormhole when computing
its hash in CYapfNodeKeyExitDir. This can cause collisions between
a position on a bridge and another one under it, but positions in
a wormhole are rare overall, having a track running along a bridge
under it is even rarer and anyway hash values need not be unique,
and this is all better than severely underusing one bit of hash
space.