Allow ships to turn under their target station sign
When multiple docks per station were introduced, the destination
tile stored for ships with an order changed from the tile in front
of the dock to the tile under the station sign, merely as an
estimate, as with trains and road vehicles. However, a check in
the pathfinding code that made ships reaching their target try to
keep sailing straight was missed in the conversion, and now ships
would not consider turning at the tile beneath the station sign if
their path happened to pass through it. This is now fixed.